Add !appIdeas to command bot. (#34)
* Add !appIdeas to command bot.

Myself and others in the community have reviewed this github link and find it very helpful.  I think it could benefit the ecosystem and honestly at least a couple times a day a new Reactiflux-er is asking "What can I build?".  This is a helpful list, non opinionated and well documented with user stories etc.  I think it would be a benefit to the Reactiflux community.

-Mark "Elsewhere" Lynch

* Sure hope this is it.

Lets cross the fingers.  Seems "Okay" Now.

* Requested white space added.

* Fixed more white space problems.

* Update features/commands.js

Co-Authored-By: Nick McCurdy <nick@nickmccurdy.com>

Co-authored-by: Nick McCurdy <nick@nickmccurdy.com>