More aggressively block crypto job posters
This adds a new system to track the user ID of anyone who posts a crypto job, and prevents them from sharing more job posts. If they continue to post (like if they’re trying to circumvent the filter), they’ll be timed out for 20 hours to prevent them posting.