Create LICENSE (#270)
* Create LICENSE

This does not change our license, as we're already using MIT in our `package.json`. Rather, it just makes it more explicit so it's easier to find via git, GitHub, and related tools.

* Update copyright year

This doesn't really matter legally, but I'm setting it to the creation date of the repository for consistency and historical reference.