Split job board cache into hiring/forhire lists
This should fix the issue we've been seeing with crashes on invalid messages. Specifically the issue was because the array that was being `shift`ed was not the same one being iterated on, thus the infinite loop.

Now the shift is a) done on the correct array, b) performed immediately after getting a reference to the underlying message, which should resolve the infinite loops we've been experiencing